+++ packages/adhere-ui-searchtable/changelog/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.0.19
+
+***
+
+2022-08-17
+
+* 加入renderSearchFormBefore和renderSearchFormAfter方法
+* onClear加入Promise
+* 加入getLimit方法
+
 # 1.0.18
 
 ***